+++ helix-core/Cargo.toml
@@ -13,6 +13,8 @@ include = ["src/**/*", "README.md"]
 [features]
 
 [dependencies]
+helix-loader = { version = "0.6", path = "../helix-loader" }
+
 ropey = "1.3"
 smallvec = "1.8"
 smartstring = "1.0.0"
@@ -33,13 +35,11 @@ toml = "0.5"
 
 similar = "2.1"
 
-etcetera = "0.3"
 encoding_rs = "0.8"
 
 chrono = { version = "0.4", default-features = false, features = ["alloc", "std"] }
 
-libloading = "0.7"
-anyhow = "1"
+etcetera = "0.3"
 
 [dev-dependencies]
 quickcheck = { version = "1", default-features = false }
