+++ helix-core/src/indent.rs
@@ -444,8 +444,6 @@ where
                 debugger: None,
                 auto_pairs: None,
             }],
-            grammar: vec![],
-            grammar_selection: None,
         });
 
         // set runtime path so we can find the queries
